Silently ignore some jobs irrelevant for admins/resellers
Test Plan: `./phpunit`

Reviewers: #hkccp_developers, mollekopf

Reviewed By: #hkccp_developers, mollekopf

Subscribers: #hkccp_developers

Differential Revision: https://git.kolab.org/D3668

+++ src/app/Jobs/User/UpdateJob.php
@@ -19,6 +19,11 @@ class UpdateJob extends UserJob
             return;
         }
 
+        if ($user->role) {
+            // Admins/resellers don't reside in LDAP (for now)
+            return;
+        }
+
         if (!$user->isLdapReady()) {
             $this->delete();
             return;
